scans CSVs now!
This is a better iteration on my previous role_counter I had in here. You provide the file path, and it tells you how many people in each role are assigned to the application by email. I just need to add in how to display the emails it doesn't find, and probably sort the roles alphabetically just for ease of finding them.